move inner message classes to top level
