+++ consensus/qbft/src/reference-test/java/org/hyperledger/besu/consensus/qbt/test/RlpTest.java
@@ -1,8 +1,23 @@
-package org.hyperledger.besu.consensus.qbt;
+/*
+ * Copyright ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.besu.consensus.qbt.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.hyperledger.besu.consensus.qbt.RlpTestCaseSpec.RlpTestInput;
+import org.hyperledger.besu.consensus.qbt.support.RlpTestCaseSpec;
+import org.hyperledger.besu.consensus.qbt.support.RlpTestInput;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
